DTLS 1.3 Update Epoch to a 64 bit counter and don't allow wrapping
This is based off a prior PR that was closed. The epoch was updated
for a 64 bit counter. A 16 bit epoch is kept when we read the
epoch from the wire in store it in the record layer.

Fixes: openssl/project#1796

+++ ssl/d1_lib.c
@@ -153,7 +153,7 @@ DTLS1_RECORD_NUMBER *dtls1_record_number_new(uint64_t epoch, uint64_t seqnum)
     return recnum;
 }
 
-void dtls1_acknowledge_sent_buffer(SSL_CONNECTION *s, uint16_t before_epoch)
+void dtls1_acknowledge_sent_buffer(SSL_CONNECTION *s, uint64_t before_epoch)
 {
     pitem *item = NULL;
     piterator iter = pqueue_iterator(&s->d1->sent_messages);
