+++ ssl/ssl_local.h
@@ -2869,7 +2869,7 @@ void dtls1_get_queue_priority(unsigned char *prio64be, unsigned short seq,
 int dtls1_retransmit_sent_messages(SSL_CONNECTION *s);
 void dtls1_clear_received_buffer(SSL_CONNECTION *s);
 void dtls1_clear_sent_buffer(SSL_CONNECTION *s, int keep_unacked_msgs);
-void dtls1_acknowledge_sent_buffer(SSL_CONNECTION *s, uint16_t before_epoch);
+void dtls1_acknowledge_sent_buffer(SSL_CONNECTION *s, uint64_t before_epoch);
 __owur OSSL_TIME dtls1_default_timeout(void);
 __owur int dtls1_get_timeout(const SSL_CONNECTION *s, OSSL_TIME *timeleft);
 __owur int dtls1_check_timeout_num(SSL_CONNECTION *s);
